Open issues for failed scheduled workflows
I noticed recently the SBOM workflow was failing and had been for a long
time and we'd not noticed. I've added issues for the other workflows
that run on a schedule and fixed the updating state.

I've also renamed the SBOM workflow to `sbom.yml` to match what it does.

Co-authored-by: Rylan Polster <rslpolster@gmail.com>