livecheck: add referer option
This adds a `referer` `url` option for `livecheck` blocks, so we can
specify a referer string that's used by curl. Similar to `user_agent`,
this value is passed to livecheck's `page_headers` and `page_content`
methods, which passes it to a related `Utils::Curl` method, and the
value is eventually used as a `curl_args` argument.

This brings livecheck requests more in line with cask requests, where
a `referer` option can be provided to the main cask `url`. This is
thankfully only required in a small number of cases but it's
necessary for any request that will not work without a referer.