Allow "all" value for gnome_devel_allowlist
The `gnome_devel_allowlist` for formulae requires a specific version
as a value (e.g., `"1.57"` to allow 1.57.x versions) and this requires
us to update the value when a new odd-numbered minor version appears.
For example, the newest `pango` version is 1.57.0 but autobump was
unable to update the formula because the `gnome_devel_allowlist`
value was `"1.55"`. `pango` doesn't use the "odd-numbered minor
versions are development releases" GNOME version scheme but we have
to manually update the allowlist because we don't have a
version-independent way of opting out of the GNOME version scheme.

This adds support for an alternative `"all"` value in
`gnome_devel_allowlist` (similar to `github_prerelease_allowlist`),
which essentially opts the formula out of the GNOME version scheme
restrictions regardless of the version.