Cask::Audit: prioritize main bundle Info.plist
`Cask::Audit.cask_bundle_min_os` checks `Info.plist` files throughout
an app but it doesn't prioritize the main `Contents/Info.plist` file,
so the audit can mistakenly return the minimum macOS version for
something else within the app. This is an issue with a recent release
of the `calibre` cask, as the main `Info.plist` file declares a
minimum macOS version of Ventura (13.3) but `brew audit` is returning
Monterey (12) from a `QtWebEngineProcess.app` in
`QtWebEngineCore.framework`. The `Frameworks` directory comes before
`Info.plist` alphabetically, so the audit never checks the main
`Info.plist` file because it already found a `min_os` value.

This addresses the issue by moving the main `Contents/Info.plist`
path (if any) to the start of `info_plist_paths`, so it's checked
before any other `Info.plist` files. This approach maintains the
existing behavior (where it can check multiple files) but it will
only fall back to checking other files if it's unable to get a
`min_os` value from the main `Info.plist` file.