chore(brew doctor): Suppress the bold disclaimer when `-q/`--quiet` is passed
The `-q`/`--quiet` flag already exists and governs the "You're ready to brew" message. I'd suggest it also govern this message, which is boilerplate.

I'm planning to use e.g. `brew doctor check_clt_up_to_date --quiet` for preflight checks in scenarios when we know we'll be installing an unbottled formula. I'd for it to either print the warning and fail or print nothing and succeed.