Add `brew lgtm` command to run multiple style checks in one command.
This is a convenience command for developers to run multiple checks in
one go.

In a tap, it will run `brew typecheck`, `brew style --changed --fix` and
`brew audit --changed --skip-style` (`--changed is newly added in this
commit`).

In a non-tap, it will run `brew typecheck`, `brew style --changed --fix`
and `brew tests --changed`.

Co-authored-by: Ruoyu Zhong <zhongruoyu@outlook.com>