shims/super/cc: use canonical_path on `<sysroot>/usr/lib`
system_library_paths is used by refurbished_args to detect
duplicate/unneeded `-L<path>` flags; however, the shim does
comparisons to `canonical_path(<path>)`. This means an SDK path like
`/Library/Developer/CommandLineTools/SDKs/MacOSX26.sdk` won't work as
it is a symlink to MacOSX26.0.sdk

This then can cause issues if build is run with `-L` to SDK path as
it can cause the wrong libraries to be linked. This is often an issue
elsewhere (e.g. seen in PHP) but this PR allows the shim to handle this
case better.