cask/artifact/app: fix up permissions for globally installed apps
Some casks (e.g. `sourcetree`) install `app` artifacts with permissions
that make them accessible only to the user that installed them. This is
problematic on multi-user systems when the app is installed into a
global location like `/Applications`.

Let's handle that by fixing up permissions appropriately when we install
an app artifact into a global location.

I've had to make this `typed: true` to match the code it inherits from
the `Moved` class, which also used `typed: true` instead of
`typed: strict`.