Shell out to `chmod` unconditionally
This is simpler and faster, since we can use `a+rX` to apply the
executable bit to directories and any files that already have any
executable permissions.

    ❯ hyperfine -w1 -C 'brew uninstall --cask --force sourcetree' -p 'git checkout cask-app-permissions' 'brew install --cask sourcetree' -n 'benchmark (v1)' -p 'git checkout cask-app-permissions-v2' -n 'bulk (v2)' 'brew install --cask sourcetree' -p 'git checkout cask-app-permissions-v3' -n 'bulk (v3)' 'brew install --cask sourcetree' --export-markdown /tmp/result-v3.md
    Benchmark 1: benchmark (v1)
      Time (mean ± σ):      4.031 s ±  0.161 s    [User: 1.842 s, System: 1.396 s]
      Range (min … max):    3.793 s …  4.413 s    10 runs

    Benchmark 2: bulk (v2)
      Time (mean ± σ):      3.987 s ±  0.088 s    [User: 1.811 s, System: 1.356 s]
      Range (min … max):    3.840 s …  4.163 s    10 runs

    Benchmark 3: bulk (v3)
      Time (mean ± σ):      3.884 s ±  0.110 s    [User: 1.771 s, System: 1.333 s]
      Range (min … max):    3.700 s …  4.014 s    10 runs

    Summary
      bulk (v3) ran
        1.03 ± 0.04 times faster than bulk (v2)
        1.04 ± 0.05 times faster than benchmark (v1)

For reference, "benchmark (v1)" is the original pure Ruby implementation
(with a fallback to shelling out when `sudo` is required). "bulk (v2)"
uses `FileUtils.chmod_R` to try to change permissions in bulk, but this
requires handling executables separately (because Ruby doesn't handle
the `+X` flag correctly).

"bulk (v3)" shells out to `chmod` unconditionally.