livecheck: add header option
This adds a `header` `url` option for `livecheck` blocks, so we can
specify headers to be used by curl. Similar to the other curl-related
options, this value is passed to livecheck's `page_headers` and
`page_content` methods, which passes it to a related `Utils::Curl`
method, and the value is eventually used as a `curl_args` argument.

This brings livecheck requests more in line with cask requests, where
a `header` option can be provided to the main cask `url` (as a string
or array of strings). There aren't any casks that use a `header`
argument but we had a need to set headers in a `livecheck` block
request years ago, so there's something to be said for making this
available.