curl_args: handle cookies argument
This adds a `cookies` parameter to `Utils::Curl.curl_args`, allowing
us to set cookies to use in a curl request. The existing `File::NULL`
approach doesn't seem to echo cookies to `stdout` like the prior
`--cookie-jar` approach, so the behavior when using a `cookies`
argument seems similar (in terms of the lack of output).