livecheck: add cookies option
This adds a `cookies` `url` option for `livecheck` blocks, so we can
specify cookies to be used by curl. Similar to the other curl-related
options, this value is passed to livecheck's `page_headers` and
`page_content` methods, which passes it to a related `Utils::Curl`
method, and the value is eventually used as a `curl_args` argument.

This brings livecheck requests more in line with cask requests, where
a `cookies` option can be provided to the main cask `url`. There are
only six casks that use a `cookies` argument but we had a need to set
cookies in a `livecheck` block request years ago, so there's something
to be said for making this available.