Cask::Audit: reduce livecheck fetches
The existing livecheck-related cask audits can end up fetching the
livecheck URL four times and this can cause problems if a server has
sensitive rate limiting. This modifies the `livecheck_version` audit
to cache the result that's also used in the `hosting_with_livecheck`
and `livecheck_https_availability` audits. This will avoid a duplicate
fetch and can avoid another if the livecheck URL uses HTTPS and the
check doesn't fail. Most casks meet that criteria, so this should be
a notable improvement.

There's still room for improvement, as the `min_os` audit calls
`cask_sparkle_min_os`, which can also fetch the livecheck URL. It
would be ideal if we could cache the livecheck content to avoid this
duplicate fetch but I don't think that's possible with the current
setup, so I'll have to revisit that later.